requirements.txt
transformers>=4.41.1
pydub>=0.25.1
openai-whisper>=20231117
python-telegram-bot>=21.2
python-telegram-bot[ext]>=21.2